+++ Package.swift
@@ -6,15 +6,11 @@ let package = Package(
         Target(
             name: "SQLite",
             dependencies: [
-                .Target(name: "SQLiteObjc"),
-                .Target(name: "CSQLite")
+                .Target(name: "SQLiteObjc")
             ]),
-        Target(
-            name: "SQLiteObjc",
-            dependencies: [
-                .Target(name: "CSQLite")
-            ]),
-        Target(
-            name: "CSQLite")
+        Target(name: "SQLiteObjc")
+    ],
+    dependencies: [
+        .Package(url: "https://github.com/jberkel/CSQLite.git", majorVersion: 0)
     ]
 )