+++ Tests/SQLiteTests/ConnectionTests.swift
@@ -5,7 +5,7 @@ import XCTest
 import sqlite3
 #elseif SQLITE_SWIFT_SQLCIPHER
 import SQLCipher
-#elseif COCOAPODS
+#else
 import CSQLite
 #endif